rename: component를 폴더 구조에 맞게 이동
