--- src/components/write-form/title.tsx
@@ -1,34 +1,0 @@
-import { WriteForm } from "@/types";
-import { UseFormReturn } from "react-hook-form";
-import {
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "../ui/form";
-import { Input } from "../ui/input";
-
-type TitleProps = {
-  form: UseFormReturn<WriteForm>;
-};
-
-export const Title = ({ form }: TitleProps) => {
-  return (
-    <>
-      <FormField
-        control={form.control}
-        name="title"
-        render={({ field }) => (
-          <FormItem>
-            <FormLabel>제목</FormLabel>
-            <FormControl>
-              <Input placeholder="제목을 입력해주세요" {...field} />
-            </FormControl>
-            <FormMessage />
-          </FormItem>
-        )}
-      />
-    </>
-  );
-};
